Move node, operator, umaryoperator to breakpoint
