+++ gb-dbg/src/debugger/breakpoints/test_parser.rs
@@ -1,8 +1,9 @@
 use super::parser::{
     address, bin_op, comb_op, expr_complete, operation, raw_register, raw_value, unary_expr,
-    unary_expr_id, Node, Operator, UnaryOperator,
+    unary_expr_id,
 };
 use crate::dbg_interfaces::CpuRegs;
+use crate::debugger::breakpoints::breakpoint::{Node, Operator, UnaryOperator};
 
 #[test]
 fn test_bin_op() {
